Replace bottomRightMenu by mobileMenu component
